Move dependency graph marshal tests
